Revert "[add] 既存 GET request ファイルで未追加 + 新規テストを integration に追加"
