--- test/common/request/get/4xx/400_13_empty_host_header_value.txt
@@ -1,4 +1,0 @@
-GET / HTTP/1.1
-Host:
-Connection: close
-
