--- test/common/request/get/4xx/400_14_transfer_encoding_and_content_length.txt
@@ -1,8 +1,0 @@
-GET /upload/test_upload_file HTTP/1.1
-Host: localhost
-Content-Length: 5
-Transfer-Encoding: chunked
-Connection: close
-Content-Type: text/plain
-
-abcde
